Windows Pageant: fix off-by-one in -c option.
Apparently I never re-tested that option when I revamped Pageant's
command-line option parsing in commit dc183e1649b429a, because it's
now off by one in figuring out which argument to treat as the start of
the command to be run.

(The new code in that commit is the same shape as the old code but
with variables renamed, and that was the mistake, because in the old
code, the argument index i pointed to the -c option, whereas in the
new code, match_opt has already advanced amo.index to the next word.
So the two index variables _shouldn't_ be treated the same.)